requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
python-telegram-bot==21.10
requests==2.32.3
sniffio==1.3.1
starlette==0.45.3
telegram==0.0.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
SQLAlchemy~=2.0.38
psycopg2==2.9.10